Fix Navbar and add Delete page
