+++ src/pages/Navbar.jsx
@@ -2,12 +2,14 @@
 import { useState } from 'react';
 import { FaBars, FaArrowLeft } from 'react-icons/fa';
 import { NavLink } from 'react-router-dom';
-// import TutorialItems  from './tutorials/TutorialItems'
+import { useSelector } from 'react-redux';
 import { signout } from '../redux/userSlice';
 import { useDispatch } from 'react-redux';
+
 const Navbar = () => {
   const [isOpen, setIsOpen] = useState(false);
   const dispatch = useDispatch();
+  const isAuthenticated = useSelector((state) => state.user.isAuthenticated);
 
   const toggleMenu = () => {
     setIsOpen(!isOpen);
@@ -15,12 +17,14 @@ const Navbar = () => {
 
   const handleSignOut = () => {
     dispatch(signout());
+  };
+
+  if (!isAuthenticated) {
+    return null;
   }
-  
 
   return (
-    ['/', '/login', '/signup'].includes(window.location.pathname) ? null : (
-      <nav className="bg-white text-[#4c4c4c] border-r border-gray-400 fixed h-full flex flex-col transition-transform duration-300 transform translate-x-0 lg:translate-x-0 lg:w-40 lg:items-center"style={{ zIndex: 1000 }}>
+<nav className="bg-white text-[#4c4c4c] border-r border-gray-400 fixed h-full flex flex-col transition-transform duration-300 transform translate-x-0 lg:translate-x-0 lg:w-80 lg:items-center" style={{ zIndex: 1000 }}>
       {/* Hamburger button for smaller screens */}
       <div className="lg:hidden cursor-pointer p-4" onClick={toggleMenu}>
         {isOpen ? <FaArrowLeft /> : <FaBars />}
@@ -32,32 +36,31 @@ const Navbar = () => {
         </li>
         <li className="p-4 cursor-pointer">
           <NavLink to="/home" className='block px-4 text-[18px] py-4 font-bold hover:bg-[#97bf0f] hover:text-white'>
-            Home
-          </NavLink>
-        </li>
-        <li className="p-4 cursor-pointer">
-          <NavLink to="/details/1" className='block px-4 text-[18px] py-4 font-bold hover:bg-[#97bf0f] hover:text-white'>
-            Details
-          </NavLink>
-        </li>
-        <li className="p-4 cursor-pointer">
-          <NavLink to="/tutorials" className='block px-4 text-[18px] py-4 font-bold hover:bg-[#97bf0f] hover:text-white'>
-            Tutorials
-            {/* <TutorialItems /> */}
+            All Tutorials
           </NavLink>
         </li>
         <li className="p-4 cursor-pointer">
           <NavLink to="/reservations" className='block px-4 text-[18px] py-4 font-bold hover:bg-[#97bf0f] hover:text-white'>
-            Reservations
+            Reserve
+          </NavLink>
+        </li>
+        <li className="p-4 cursor-pointer">
+          <NavLink to="/myreservations" className='block px-4 text-[18px] py-4 font-bold hover:bg-[#97bf0f] hover:text-white'>
+            My Reservations
           </NavLink>
         </li>
         <li className="p-4 cursor-pointer">
           <NavLink to="/classes" className='block px-4 text-[18px] py-4 font-bold hover:bg-[#97bf0f] hover:text-white'>
-            Add class
+            Add Tutorial
+          </NavLink>
+        </li>
+        <li className="p-4 cursor-pointer">
+          <NavLink to="/delete" className='block px-4 text-[18px] py-4 font-bold hover:bg-[#97bf0f] hover:text-white'>
+            Delete Tutorial
           </NavLink>
         </li>
         <li>
-          <NavLink to="/" onClick={() => handleSignOut()} className='block px-4 text-[18px] py-4 font-bold hover:bg-[#97bf0f] hover:text-white'>
+          <NavLink to="/" onClick={() => handleSignOut()} className='block px-4 text-[18px] py-4 mt-28 ml-auto font-bold hover:bg-[#97bf0f] hover:text-white'>
             Logout
           </NavLink>
         </li>
@@ -87,7 +90,6 @@ const Navbar = () => {
         </small>
       </div>
     </nav>
-    )
   );
 };
